+++ resources/views/cart/show.blade.php
@@ -1,3 +1,3 @@
-<x-layouts::storefront :title="__('My Cart')">
+<x-layouts::storefront :title="__('My Cart')" :show-announcements="false">
     <livewire:storefront.cart-page />
 </x-layouts::storefront>