+++ resources/views/checkout/show.blade.php
@@ -1,3 +1,3 @@
-<x-layouts::storefront :title="__('Checkout')">
+<x-layouts::storefront :title="__('Checkout')" :show-announcements="false">
     <livewire:storefront.checkout-page />
 </x-layouts::storefront>